+++ backend/windmill-common/src/variables.rs
@@ -6,6 +6,7 @@
  * LICENSE-AGPL for a copy of the license.
  */
 
+use chrono::Utc;
 use magic_crypt::{MagicCrypt256, MagicCryptError, MagicCryptTrait};
 use serde::{Deserialize, Serialize};
 
@@ -147,6 +148,8 @@ pub async fn decrypt_value_with_mc(
     })?)
 }
 
+pub const WM_SCHEDULED_FOR: &str = "WM_SCHEDULED_FOR";
+
 pub async fn get_reserved_variables(
     db: &DB,
     w_id: &str,
@@ -162,6 +165,7 @@ pub async fn get_reserved_variables(
     step_id: Option<String>,
     root_flow_id: Option<String>,
     jwt_token: Option<String>,
+    scheduled_for: Option<chrono::DateTime<Utc>>,
 ) -> Vec<ContextualVariable> {
     let state_path = {
         let trigger = if schedule_path.is_some() {
@@ -245,6 +249,14 @@ pub async fn get_reserved_variables(
             name: "WM_JOB_ID".to_string(),
             value: job_id.to_string(),
             description: "Job id of the current script".to_string(),
+            is_custom: false,
+        },
+        ContextualVariable {
+            name: WM_SCHEDULED_FOR.to_string(),
+            value: scheduled_for
+                .map(|ts| ts.to_string())
+                .unwrap_or_else(|| "".to_string()),
+            description: "date-time in UTC (e.g: 2014-11-28T12:45:59.324310806Z) of when the job was scheduled".to_string(),
             is_custom: false,
         },
         ContextualVariable {