+++ frontend/src/lib/common.ts
@@ -26,9 +26,10 @@ export interface SchemaProperty {
 	contentEncoding?: 'base64' | 'binary'
 	format?: string
 	items?: {
-		type?: 'string' | 'number' | 'bytes' | 'object'
+		type?: 'string' | 'number' | 'bytes' | 'object' | 'resource'
 		contentEncoding?: 'base64'
 		enum?: string[]
+		resourceType?: string
 	}
 	min?: number
 	max?: number