+++ frontend/src/lib/components/ArgInput.svelte
@@ -30,6 +30,8 @@
 	import SchemaFormDnd from './schema/SchemaFormDND.svelte'
 	import SchemaForm from './SchemaForm.svelte'
 	import { deepEqual } from 'fast-equals'
+	import DynSelect from './DynSelect.svelte'
+	import type { Script } from '$lib/gen'
 
 	export let label: string = ''
 	export let value: any
@@ -47,12 +49,14 @@
 	export let disabled = false
 	export let itemsType:
 		| {
-				type?: 'string' | 'number' | 'bytes' | 'object'
+				type?: 'string' | 'number' | 'bytes' | 'object' | 'resource'
 				contentEncoding?: 'base64'
 				enum?: string[]
 				multiselect?: string[]
+				resourceType?: string
 		  }
 		| undefined = undefined
+
 	export let displayHeader = true
 	export let properties: { [name: string]: SchemaProperty } | undefined = undefined
 	export let nestedRequired: string[] | undefined = undefined
@@ -79,6 +83,11 @@
 	export let editor: SimpleEditor | undefined = undefined
 	export let orderEditable = false
 	export let shouldDispatchChanges: boolean = false
+	export let helperScript:
+		| { type: 'inline'; path?: string; lang: Script['language']; code: string }
+		| { type: 'hash'; hash: string }
+		| undefined = undefined
+	export let otherArgs: Record<string, any> = {}
 
 	let oneOfSelected: string | undefined = undefined
 	async function updateOneOfSelected(oneOf: SchemaProperty[] | undefined) {
@@ -372,7 +381,7 @@
 															on:change={(x) => fileChanged(x, (val) => (value[i] = val))}
 															multiple={false}
 														/>
-													{:else if itemsType?.type == 'object'}
+													{:else if itemsType?.type == 'object' && itemsType?.resourceType === undefined}
 														<JsonEditor code={JSON.stringify(v, null, 2)} bind:value={v} />
 													{:else if Array.isArray(itemsType?.enum)}
 														<ArgEnum
@@ -392,6 +401,8 @@
 															enum_={itemsType?.enum ?? []}
 															enumLabels={extra['enumLabels']}
 														/>
+													{:else if itemsType?.type == 'resource' && itemsType?.resourceType}
+														<ResourcePicker bind:value={v} resourceType={itemsType?.resourceType} />
 													{:else}
 														<input type="text" bind:value={v} id="arg-input-array" />
 													{/if}
@@ -450,6 +461,14 @@
 						/>
 					</div>
 				</div>
+			{:else if inputCat == 'dynselect'}
+				<DynSelect
+					name={label}
+					args={otherArgs}
+					{helperScript}
+					bind:value
+					entrypoint={format.substring('dynselect_'.length)}
+				/>
 			{:else if inputCat == 'resource-object' && resourceTypes == undefined}
 				<span class="text-2xs text-tertiary">Loading resource types...</span>
 			{:else if inputCat == 'resource-object' && (resourceTypes == undefined || (format.split('-').length > 1 && resourceTypes.includes(format.substring('resource-'.length))))}