+++ frontend/src/lib/components/ScheduleEditorInner.svelte
@@ -22,6 +22,7 @@
 	} from '$lib/gen'
 	import { enterpriseLicense, userStore, workspaceStore } from '$lib/stores'
 	import { canWrite, emptyString, formatCron, sendUserToast } from '$lib/utils'
+	import { base } from '$lib/base'
 	import { createEventDispatcher } from 'svelte'
 	import Section from '$lib/components/Section.svelte'
 	import { List, Loader2, Save } from 'lucide-svelte'
@@ -417,7 +418,7 @@
 						variant="border"
 						startIcon={{ icon: List }}
 						disabled={!allowSchedule || pathError != '' || emptyString(script_path)}
-						href={`/runs/${script_path}?show_schedules=true&show_future_jobs=true`}
+						href={`${base}/runs/${script_path}?show_schedules=true&show_future_jobs=true`}
 					>
 						View runs
 					</Button>