+++ frontend/src/routes/(root)/(logged)/+layout.svelte
@@ -28,7 +28,8 @@
 		hubBaseUrlStore
 	} from '$lib/stores'
 	import CenteredModal from '$lib/components/CenteredModal.svelte'
-	import { afterNavigate, beforeNavigate, goto } from '$app/navigation'
+	import { afterNavigate, beforeNavigate } from '$app/navigation'
+	import { goto } from '$lib/navigation'
 	import UserSettings from '$lib/components/UserSettings.svelte'
 	import SuperadminSettings from '$lib/components/SuperadminSettings.svelte'
 	import WindmillIcon from '$lib/components/icons/WindmillIcon.svelte'
@@ -45,6 +46,7 @@
 	import GlobalSearchModal from '$lib/components/search/GlobalSearchModal.svelte'
 	import MenuButton from '$lib/components/sidebar/MenuButton.svelte'
 	import { setContext } from 'svelte'
+	import { base } from '$app/paths'
 
 	OpenAPI.WITH_CREDENTIALS = true
 	let menuOpen = false
@@ -158,22 +160,22 @@
 		favoriteLinks = [
 			...scripts.map((s) => ({
 				label: s.summary || s.path,
-				href: `/scripts/get/${s.hash}`,
+				href: `${base}/scripts/get/${s.hash}`,
 				kind: 'script' as 'script'
 			})),
 			...flows.map((f) => ({
 				label: f.summary || f.path,
-				href: `/flows/get/${f.path}`,
+				href: `${base}/flows/get/${f.path}`,
 				kind: 'flow' as 'flow'
 			})),
 			...apps.map((f) => ({
 				label: f.summary || f.path,
-				href: `/apps/get/${f.path}`,
+				href: `${base}/apps/get/${f.path}`,
 				kind: 'app' as 'app'
 			})),
 			...raw_apps.map((f) => ({
 				label: f.summary || f.path,
-				href: `/apps/get_raw/${f.version}/${f.path}`,
+				href: `${base}/apps/get_raw/${f.version}/${f.path}`,
 				kind: 'raw_app' as 'raw_app'
 			}))
 		]
@@ -182,11 +184,11 @@
 	function pathInAppMode(pathname: string | undefined): boolean {
 		if (!pathname) return false
 		return (
-			pathname.startsWith('/apps') ||
-			pathname.startsWith('/flows/add') ||
-			pathname.startsWith('/flows/edit') ||
-			pathname.startsWith('/scripts/add') ||
-			pathname.startsWith('/scripts/edit')
+			pathname.startsWith(base + '/apps') ||
+			pathname.startsWith(base + '/flows/add') ||
+			pathname.startsWith(base + '/flows/edit') ||
+			pathname.startsWith(base + '/scripts/add') ||
+			pathname.startsWith(base + '/scripts/edit')
 		)
 	}
 	afterNavigate((n) => {
@@ -203,7 +205,7 @@
 		}
 	}
 
-	let devOnly = $page.url.pathname.startsWith('/scripts/dev')
+	let devOnly = $page.url.pathname.startsWith(base + '/scripts/dev')
 
 	async function loadCopilot(workspace: string) {
 		workspacedOpenai.init(workspace)
@@ -253,7 +255,7 @@
 		globalSearchModal?.openSearchWithPrefilledText(text)
 	}
 
-	setContext("openSearchWithPrefilledText", openSearchModal)
+	setContext('openSearchWithPrefilledText', openSearchModal)
 </script>
 
 <svelte:window bind:innerWidth />